feat: add map with custom markers for user positions in GroupDetails
